requirements.txt
# Auto-generated from current environment's `pip list`.
# Note: This is a snapshot and may include system/driver packages.
# For reproducible setups, prefer `pip freeze` within the same environment.

absl-py==2.3.1
accelerate==1.11.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.1.2
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.25.1
docstring_parser==0.17.0
einops==0.8.1
embreex==2.17.7.post7
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
gsplat==1.5.3+pt23cu118
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.1
idna==3.11
imageio==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.3.0
jaxtyping==0.2.29
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
lpips==0.1.4
lxml==6.0.2
manifold3d==3.3.2
mapbox_earcut==1.0.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
modelscope==1.32.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
nerfview==0.0.2
networkx==3.3
ninja==1.13.0
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.82
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
nvitop==1.6.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
peft==0.9.0
pillow==11.3.0
pip==25.3
platformdirs==4.5.0
pluggy==1.6.0
protobuf==6.33.1
psutil==7.1.3
pycollada==0.9.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyliblzfse==0.4.1
pyparsing==3.2.5
pypose==0.7.3
PySocks==1.7.1
pytest==9.0.1
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.44.0
setuptools==80.9.0
shapely==2.1.2
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
splines==0.3.3
svg.path==7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorly==0.9.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.15.2
tomli==2.3.0
torch==2.1.2
torch-fidelity==0.3.0
torchmetrics==1.8.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.38.0
trimesh==4.9.0
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tyro==0.9.1
tzdata==2025.2
urllib3==2.5.0
vhacdx==0.0.9
viser==0.2.1
wadler_lindig==0.1.7
wandb==0.23.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.23.post1
xxhash==3.6.0
yourdfpy==0.0.58
zipp==3.23.0

# (removed) pycolmap for data parsing
# git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e

# (optional) nerfacc for torch version rasterization
# git+https://github.com/nerfstudio-project/nerfacc

viser
nerfview==0.0.2
imageio[ffmpeg]
numpy<2.0.0
scikit-learn
tqdm
torchmetrics[image]
opencv-python
tyro>=0.8.8
Pillow
tensorboard
pyyaml
seaborn
splines
tensorly
pypose
matplotlib
#git+https://ghproxy.com/https://github.com/rahul-goel/fused-ssim@1272e21a282342e89537159e4bad508b19b34157
git+https://ghproxy.com/https://github.com/nerfstudio-project/gsplat@dd66cbd597f376f103e90de8b6265087b482eac1
